Reset the application state completely on back
This is the simplest possible answer to the lingering issues
with bugs around the back button. I've replaced the back button
with a 'Home' button on Distribute and Export, and completely
wipe the state when the Home screen loads. We might want to add
a confirm step so users don't lose data accidentally.