Creates Vagrantfile with ubuntu/trusty64
Trusty dusty 64, good enough for a start. The 14.04 LTS release is
showing its age, but it's mostly useful for the automatic mounting of
the /vagrant directory, which is currently broken in the 16.04 vagrant
image.

Ignores .vagrant directory, since it's only used for local VMs,
and should never be committed to version control.