Use semver to compare versions
Adds the semver library to perform the comparison. If either the "latest
release" version fetched from GitHub or the currentVersion of the
running app is invalid according to the semver library, an error message
is logged to the console, but no message is displayed to the end user.

This behavior should allow us to recover from any mistakes in setting
the latest release tag seamlessly and without needlessly alarming end
users.