Creates pardir for app icons under Linux
The previous logic was naively assuming that the parent directory
existed. Bizarrely the `icns2png` command *fails silently* if it cannot
write the output files. :table_flip:

Added a new tasknto create the parent directory, and modified the
icns2png conversion command to catch errors if encountered again in the future.

Closes #32.