Add CircleCI status badge to README
Since Sunder is still a private repository, I had to create a CircleCI
[API token][0] in order to get a working status badge. This token is
scoped to the `freedomofpress/sunder` repository and the `status` scope.
After I created the token, CircleCI automatically started embedding it
in the status badge URL it provides.

When we make the Sunder repo public, we should replace this URL with the
public equivalent (which doesn't need an API token) and destroy this API
token through the CircleCI interface.

[0]: https://circleci.com/gh/freedomofpress/sunder/edit#api