Make before/after async/await again
The Chai test runner's before/after hooks were originally async, but I
made them sync while I was initially working on this PR to reduce
potential sources of confusion. Now that the dust has settled from
integrating Spectron, it's clear that these can be made async/await
again, which is more consistent with the code in the rest of the e2e
tests.