Fix e2e test
Now that we've integrated Spectron, we can access Electron API's
directly from within the e2e tests, which makes testing
clipboard-related functionality straightforward.

I added the className to CopyButton and unique id's to each ShareRow to
make it easier to construct an explicit WebdriverIO selector. I think
explicit selectors (and correspondingly, careful design of classes and
ID's) are good because they make tests easier to read and reason about,
and reduce the likelihood of test failures due to imprecise selectors.