Store and use directory that a shard was saved to.
This changes the behavior of the distribute screen save file dialog, so
that it reopens in the last directory that a shard was saved to,
instead of a random default.

Doing this manually instead of letting the OS do it for us (by passing
`undefined` to defaultPath), lets us still specify a suggested file
name. This is seems to be a limitation of the Electron API.