Version control package-lock.json
I just upgraded Node and NPM to the latest LTS to confirm that Sunder
works with them. NPM 5 introduced package-lock.json, which is supposed
to be committed to version control. I ran a fresh node install and
tested that most things are working, so I think the set of dependencies
currently in the lock files is good to commit.