Upgrade CircleCI Node
I've been using Node v8.x for all of my recent development, and just
updated the README to reflect this assumption about the development
environment. We should always try to align our CI environment with our
development environment as much as possible so CI outcomes are
concordant with local outcomes.

As far as I know, we are still using CircleCI's 1.0 platform with the
Ubuntu 14.04 build image [0], so the most recent version of Node we can
use on CircleCI is v8.2.0. It would be better to use the same release we
are using in development (the current LTS, v8.9.4), but this is the best
we can do given the circumstances and I think it's still a worthwhile
change.

[0]: https://circleci.com/docs/1.0/build-image-trusty/#nodejs