Creates Dockerfile for sunder-build image
The "sunder-build" image contains the necessary tooling for kicking off
a build of the Sunder Electron app, creating artifacts in the form of
Debian packages.

The apt package installation logic, pulling in the dependencies, was
ported from the existing Ansible config. The `npm` build commands will
be placed in an adjacent script, since running those requires the
application code from the git repo to be mounted inside the container,
which will happen as a container "run" action, not "build" action.