Improve docker build logic for user over-ride
Previous logic did not take into account a condition where a user wants
to run the container with a volume mapped to a uid which doesnt already
exist in the container. Now we are explicitly setting the `node` user to
match the builder's user id.

Also introduce some logic to correct volume mounting permissions from a
named volume mounted over-top of node_modules