Update from moment 2.19.2
Moment is affected by CVE2017-18214 (DoS via crafted date string). Not only is it a dev dependency, sunder does not use this functionality and therefore shouldn't be affected. To avoid receiving and ignoring security notifications, version bump of concurrently is non-trivial and requires code changes. Overriding and bumping version of moment in `package.json` appears to be an easy fix.