Installs node_modules as dependency if missing
The `npm install` task takes ~23s even when no changes are required, so
let's not delay the quick test runs (e.g. 3s for `make test-unit`) with
unnecessary npm calls. Instead, we'll check for existence of the
`node_modules/` directory locally, and install the npm modules only if
it's missing. That'll still support automatic integration with the `make
clean` target for clean builds.