Force HTTPS fetching of rust cargo crates
As part of the Circle CI functionality, a custom ~/.gitconfig is written
that forces git to use SSH URLs, rather than HTTPS URLs, for github.com.
That's fine and dandy for cloning the repo being tested, but will break
cloning of repos for e.g. Rust crates.

It's a bit of a hack, but after we've successfully bootstrapped the repo
to be tested, we can simply remove the custom ~/.gitconfig and restore
normal HTTPS functionality.

See https://github.com/rust-lang/cargo/issues/2078 for more detail.