FileInput passes an array to onChange
This generalizes FileInput so that it always passes an array of one or
more files to it's onChange property, rather than just the attributes of
a single file.

The onChange listeners are updated to handle an array instead of a
single file.

This is another incremental step towards handling multiple shares
selected at the same time.