Remove cryptiles dependency
A security vulnerability was found in cryptiles, a deeply nested
dependency in our package-lock.json. Updating spectron, fsevents, and
request modules to their latest versions effectively removes this
dependency.

Also worth noting is that these are all tools for development and test,
so Sunder should be unaffected by the vulnerability itself.