Fix hot-reloading
This localhost endpoint is white listed because in HOT mode this is
where the bundle.js is initially loaded from, however, hot updates are
applied via a <hash>.hot-update.js file from the same local web server,
so in order to enable them we can simply whitelist all scripts from this
endpoint.

Strictly speaking, it's not good practice to have this sort of hole in
the CSP, but there's no other easy way to support live reloading of
resources with webpack's granlarity. Ideally, we would use a less
lenient CSP when building and packaging for distribution.